Add partial support for data ranges
This patch adds partial support for data ranges specified by WPS processes.  Adds two fields to hold range min and max values, and populates them if the user specifies a numeric input (real or integer), along with a value that contains a "-".  Without this patch, that combination will cause a crash.

I suspect that this method of expressing a data range is not fully standards compliant (though I am not sure), but there are WPS servers in the wild that specify data this way.  My thinking is that where the server's intent is clear, it is better to read the data and get on with life than it is to crash.